TASK7 LATEST CORRECTION

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -3,7 +3,7 @@
 DELIMITER $$
 CREATE PROCEDURE ComputeAverageScoreForUser (IN user_id INT)
 BEGIN
-	DECLARE aver_score INT;
+	DECLARE aver_score DECIMAL(10, 2);
 	SELECT AVG(score) INTO aver_score FROM corrections AS C WHERE C.user_id IN (user_id);
 	UPDATE users SET average_score = aver_score WHERE id = user_id;
 END $$
